+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fix comparison order in `compareBounds`
 * removed `compareBounds`' forced restriction to `Levitated`
 * `Data`, `Typeable`, `Generic` instances for `Interval`
+* `Data` instance for `Adjacency`
 
 ## 0.3.1.0 - 2022-07-13