+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * `Data` instance for `Adjacency`
 * `imin`, `iinf`, `isup`, `imax` no longer return `Bound`s
 * fix `difference` in cases `MetBy` and `After`
+* fix regression in smart constructor ordering
 
 ## 0.3.1.0 - 2022-07-13