+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * fix `difference` in cases `MetBy` and `After`
 * fix regression in smart constructor ordering
 * add example to `measuring`
+* improve implementation of `unionsAsc`
 
 ## 0.3.1.0 - 2022-07-13